+++ inbox/init.lua
@@ -1,5 +1,5 @@
 
-local S = homedecor_i18n.gettext
+local S = homedecor.gettext
 
 local inbox = {}
 local screwdriver = rawget(_G, "screwdriver") or {}